requirements.txt
streamlit
speechmatics-python
yt-dlp
setuptools  # if using Python 3.12
python-decouple